style: 전체적인 코드 스타일 수정
