+++ src/main/java/domain/RacingGame.java
@@ -1,41 +1,31 @@
 package domain;
 
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 
 public class RacingGame {
 
-    private final List<Car> cars;
+    private final CarGroup carGroup;
     private final RandomNumberGenerator randomNumberGenerator;
 
     public RacingGame(String[] names) {
-        this.cars = new ArrayList<>();
-        setUp(names);
+        this.carGroup = new CarGroup(names);
         this.randomNumberGenerator = new RandomNumberGenerator();
     }
 
     //TODO: 테스트
     public void race() {
-        for (Car car : cars) {
+        for (Car car : carGroup.getCars()) {
             car.move(randomNumberGenerator);
         }
     }
 
     public RacingResult getRacingResult(){
         Map<String, Integer> history = new LinkedHashMap<>();
-        for (Car car : cars) {
+        for (Car car : carGroup.getCars()) {
             history.put(car.getName(), car.getPosition());
         }
 
         return new RacingResult(history);
-    }
-
-    private void setUp(String[] names) {
-        for (String name : names) {
-            Car car = new Car(name);
-            cars.add(car);
-        }
     }
 }
